app/requirements.txt
numpy==1.26.4
tensorflow==2.17.0
psycopg2-binary==2.9.9
pillow==10.3.0